[DistDataloader] Update implementation, add nested.py
